requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Configuration and Environment
pyyaml==6.0.1
python-dotenv==1.0.0

# HTTP Client and WebSocket
httpx==0.25.2
websockets==12.0

# AI and LLM Integration
openai==1.3.7
anthropic==0.7.7
tiktoken==0.5.1

# Memory and Vector Storage
chromadb==0.4.17
sentence-transformers==2.2.2
numpy==1.24.4

# Async and Concurrency
asyncio-throttle==1.0.2
aiofiles==23.2.1

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Security and Authentication
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
click==8.1.7
rich==13.7.0
jinja2==3.1.2
